Implement mocriphone dropdown selection control.
