+++ osu.Game.Rulesets.Karaoke/Screens/Config/Sections/Input/MicrophoneSettings.cs
@@ -3,10 +3,6 @@
 
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
-using osu.Framework.Input;
-using osu.Framework.Localisation;
-using osu.Game.Graphics.UserInterface;
-using osu.Game.Overlays.Settings;
 using osu.Game.Rulesets.Karaoke.Configuration;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Config.Sections.Input
@@ -18,27 +14,14 @@ namespace osu.Game.Rulesets.Karaoke.Screens.Config.Sections.Input
         [BackgroundDependencyLoader]
         private void load()
         {
-            var microphoneManager = new MicrophoneManager();
             Children = new Drawable[]
             {
-                new MicrophoneDeviceSettingsDropdown
+                new SettingsMicrophoneDeviceDropdown
                 {
                     LabelText = "Microphone devices",
-                    Items = microphoneManager.MicrophoneDeviceNames,
                     Current = Config.GetBindable<string>(KaraokeRulesetSetting.MicrophoneDevice)
                 },
             };
-        }
-
-        private class MicrophoneDeviceSettingsDropdown : SettingsDropdown<string>
-        {
-            protected override OsuDropdown<string> CreateDropdown() => new MicrophoneDeviceDropdownControl();
-
-            private class MicrophoneDeviceDropdownControl : DropdownControl
-            {
-                protected override LocalisableString GenerateItemText(string item)
-                    => string.IsNullOrEmpty(item) ? "Default" : base.GenerateItemText(item);
-            }
         }
     }
 }
